refactor: separation interfeces (DD-2)
