--- src/app/features/auth/interfaces/auth.interface.ts
@@ -1,11 +1,0 @@
-export interface UserRequest {
-  email: string;
-  password: string;
-  confirmPassword: string;
-  name: string;
-  lastName: string;
-}
-
-export interface UserResponse {
-  token: string;
-}
